New file for DB functions Created DB table Basic insert statement
